Initial Commit of machine/machined CLI and REST api service
There are two binaries, machine and machined.  The former is the client
which will interact with machined over the UNIX socket on which it
listens for API requests.

machined runs a gin-gonic/gin web service framework for handling the
REST api for interacting with the machine service.

machine is a client using the resty http client with a custom transport
to function over UNIX socket.