Implement run via init+start
Move the logic in the run command which was starting up an
already-defined cluster into start and export the general logic
of starting a cluster by name.  Refactor init code and export the
general logic of creating a cluster from config so that it can be
reused.

Run now reads the cluster name and config args, then calls the two
functions: create cluster, start cluster