Use IDE for cdroms to be more friendly to legacy ISO images
Generally we'd prefer to use virtio-blk for better performance with
ISOs but some code may expect isos to be in the /dev/cdrom path which
isn't the case with virtio-blk as cdrom devices.  Go ahead and use a
legacy AHCI controller and ide-cd device.

- Updated to qcli v0.0.6 release to pickup IDEControllerDevices in the
Config object, and BootIndex as strings
- Remove some cdrom debug messages
- Fix up the BlockDevice and Bus attach code for IDE CDROMs