bootindex: once more into the breach dear friends
OK.  QEMU uses zero-based bootindex values, and using an int for the
type of the BootIndex field made it difficult to distinguished between
unset and the value of zero.... yeah.

Let's go back to strings for BootIndex as a string type which is easy
to see if it's unset (empty) and then use strconv/sprintf to switch
between string and int as needed.